refactor(hosting): tighten startup boundary ownership
Why:
- Ultra-review found remaining coupling between command registration, parsing metadata, and hidden supervisor invocation validation.

What Changed:
- Split side-effect-free CLI command metadata from ConsoleAppFramework registration and moved command-token classification out of parsing policy ownership.
- Preserved test command framework pass-through while keeping daemon/logs/ops/skills/query pre-dispatch JSON errors centralized.
- Model hidden supervisor invocation validity explicitly and centralize the hidden supervisor argument sequence used by launchers and parser.

Impact:
- Public CLI command names, JSON envelopes, and hidden supervisor invalid-invocation exit behavior remain unchanged.
