+++ src/Ucli/Hosting/Cli/Common/Parsing/CliParseErrorJsonPolicy.cs
@@ -1,7 +1,7 @@
 using ConsoleAppFramework;
+using MackySoft.Ucli.Hosting.Cli.Common.Catalog;
 using MackySoft.Ucli.Hosting.Cli.Common.Contracts;
 using MackySoft.Ucli.Hosting.Cli.Common.Execution;
-using MackySoft.Ucli.Hosting.Cli.Common.Startup;
 
 namespace MackySoft.Ucli.Hosting.Cli.Common.Parsing;
 
@@ -21,24 +21,21 @@ internal static class CliParseErrorJsonPolicy
         };
     }
 
-    /// <summary> Emits a JSON invalid-argument result when framework parsing failed before a command handler started. </summary>
+    /// <summary> Creates a JSON invalid-argument result when framework parsing failed before a command handler started. </summary>
     /// <param name="args"> The command-line arguments passed to the process. </param>
-    /// <returns> <see langword="true" /> when a JSON result was emitted. </returns>
+    /// <returns> The parse-error result to emit, or <see langword="null" /> when no parse failure was captured. </returns>
     /// <exception cref="ArgumentNullException"> Thrown when <paramref name="args" /> is <see langword="null" />. </exception>
-    public static bool TryEmitParseErrorResult (string[] args)
+    public static CommandResult? TryCreateParseErrorResult (string[] args)
     {
         ArgumentNullException.ThrowIfNull(args);
 
         if (CommandExecutionState.HasStarted || !ParseErrorBuffer.HasAny)
         {
-            return false;
+            return null;
         }
 
-        var commandName = UcliCommandCatalog.ResolveResultCommandName(args);
+        var commandName = UcliCommandMetadataCatalog.ResolveResultCommandName(args);
         var parseErrorMessage = string.Join(" ", ParseErrorBuffer.Messages);
-        var parseErrorResult = CommandResult.InvalidArgument(commandName, parseErrorMessage);
-        CommandResultWriter.WriteToStandardOutput(parseErrorResult);
-        Environment.ExitCode = parseErrorResult.ExitCode;
-        return true;
+        return CommandResult.InvalidArgument(commandName, parseErrorMessage);
     }
 }
