+++ tests/Ucli.Tests/Hosting/Cli/CommandTokenClassifierTests.cs
@@ -1,4 +1,4 @@
-using MackySoft.Ucli.Hosting.Cli.Common.Parsing;
+using MackySoft.Ucli.Hosting.Cli.Common.Tokens;
 
 namespace MackySoft.Ucli.Tests;
